requirements.txt
astroid==2.3.3
click==7.1.1
cycler==0.10.0
gitdb==4.0.2
GitPython==3.1.0
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.2.0
mccabe==0.6.1
numpy==1.18.1
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
schedule==0.6.0
six==1.14.0
smmap==3.0.1
tabulate==0.8.6
typed-ast==1.4.1
wrapt==1.11.2
